Split elementwise functions into two more files and added them to the build
